feat: add basic auth like login, logout, register by username + password, gen token, refresh-token and send email verify

+++ src/main/java/com/example/schedulemeetingbe/entity/RefreshToken.java
@@ -17,7 +17,7 @@ public class RefreshToken {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "refresh_token_id")
-    private Integer refreshId;
+    private Long refreshId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
